feat(poc): integrate PoC requests into vLLM scheduler for automatic batching
Replace collective_rpc bypass with native scheduler integration for PoC
requests. This enables automatic batching within token budget and allows
PoC to coexist with chat requests through the same scheduling path.

Changes:
- Add poc_compute() method to AsyncLLMEngine and MQLLMEngineClient
- Update /generate endpoint to submit individual nonces via scheduler
- Remove _generate_worker_loop and manual batching code from routes.py
- Skip PoC sequences in _schedule_running (prefill-only, no decode)
- Bypass KV cache allocation for PoC: _can_append_slots, _append_slots
- Skip PoC in sampling metadata preparation (no sampling needed)
- Handle PoC outputs before sampler output indexing in _process_model_outputs
- Add uses_prompt_embeds() check for PoC to prevent mixed batching
- Update poc_smoke_test.py to use LLMEngine.add_request() directly
- Update poc_e2e_test.py to use new scheduler-based /generate endpoint

The batch_size parameter is now obsolete - scheduler auto-batches PoC
requests based on available token budget alongside regular chat requests.

+++ vllm/poc/gpu_random.py
@@ -228,7 +228,8 @@ def generate_haar_orthogonal_matrices(
         seed = _seed_from_string(
             f"{block_hash}_{public_key}_nonce_{nonce}_haar_qr_{k}"
         )
-        A = _normal(seed, k * k, device).view(k, k).to(dtype)
+        # QR decomposition requires float32, convert to target dtype after.
+        A = _normal(seed, k * k, device).view(k, k).to(torch.float32)
 
         Q, R = torch.linalg.qr(A, mode="reduced")
 
@@ -238,6 +239,6 @@ def generate_haar_orthogonal_matrices(
         signs = torch.where(signs == 0, torch.ones_like(signs), signs)
         Q = Q * signs.unsqueeze(0)
 
-        Qs.append(Q)
+        Qs.append(Q.to(dtype))
 
     return torch.stack(Qs, dim=0)
